requirements.txt
google-cloud-bigquery==3.17.2
python-dotenv==1.0.1
weaviate-client==4.5.0